Fix crash on setblock
There was no issue made about this, but it crashes on setblock (For example, the_aether:blue_portal when there is no glowstone below it). Previously to this update, a StackOverflowError would occur. Now, it will stop looping when the Y value is below zero, since there are no more blocks below and thus no reason to continue looping into the void.